代码生成功能更新

+++ kinit_fast_task/scripts/app_generate/v1/generate_crud.py
@@ -20,20 +20,24 @@ class CrudGenerate(GenerateBase):
         self.task_log = task_log
         self.task_log.info("开始生成 CRUD 代码, CRUD 文件地址为：", self.file_path, is_verbose=True)
 
-    def write_generate_code(self):
+    def write_generate_code(self, overwrite: bool = False):
         """
         生成 crud 文件，以及代码内容
         """
         if self.file_path.exists():
-            self.task_log.warning("CRUD 文件已存在，正在删除重新写入")
-            self.file_path.unlink()
+            if overwrite:
+                self.task_log.warning("CRUD 文件已存在, 已选择覆盖, 正在删除重新写入.")
+                self.file_path.unlink()
+            else:
+                self.task_log.warning("CRUD 文件已存在, 未选择覆盖, 不再进行 CRUD 代码生成.")
+                return False
 
         self.file_path.parent.mkdir(parents=True, exist_ok=True)
         self.file_path.touch()
 
         code = self.generate_code()
         self.file_path.write_text(code, "utf-8")
-        self.task_log.success("CRUD 代码创建完成")
+        self.task_log.success("CRUD 代码写入完成")
 
     def generate_code(self):
         """
